adds tmux support
